+++ core/src/main/scala/org/apache/spark/sql/crossdata/execution/datasources/XDDdlParser.scala
@@ -17,6 +17,7 @@ package org.apache.spark.sql.crossdata.execution.datasources
 
 import java.util.UUID
 
+import org.apache.spark.sql.catalyst.TableIdentifier
 import org.apache.spark.sql.catalyst.analysis.UnresolvedRelation
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
 import org.apache.spark.sql.crossdata.XDContext
@@ -38,6 +39,10 @@ class XDDdlParser(parseQuery: String => LogicalPlan, xDContext: XDContext) exten
   protected val INSERT = Keyword("INSERT")
   protected val INTO = Keyword("INTO")
   protected val VALUES = Keyword("VALUES")
+  protected val GLOBAL = Keyword("GLOBAL")
+  protected val INDEX = Keyword("INDEX")
+  protected val ON = Keyword("ON")
+  protected val PK = Keyword("PK")
   //Streaming keywords
   protected val EPHEMERAL = Keyword("EPHEMERAL")
   protected val SHOW = Keyword("SHOW")
@@ -62,8 +67,7 @@ class XDDdlParser(parseQuery: String => LogicalPlan, xDContext: XDContext) exten
   override protected lazy val ddl: Parser[LogicalPlan] =
 
     createTable | describeTable | refreshTable | importStart | dropTable | dropExternalTable |
-      createView | createExternalTable | dropView | addJar | streamingSentences | insertIntoTable | addApp | executeApp
-
+      createView | createExternalTable | dropView | addJar | streamingSentences | insertIntoTable | addApp | executeApp | createGlobalIndex
 
   protected lazy val streamingSentences: Parser[LogicalPlan] =
     describeEphemeralTable | showEphemeralTables | createEphemeralTable | dropAllEphemeralQueries  | dropAllEphemeralTables |
@@ -101,7 +105,6 @@ class XDDdlParser(parseQuery: String => LogicalPlan, xDContext: XDContext) exten
 
   protected lazy val arrayValues: Parser[Any] =
     ("[" ~> repsep(mapValues | token, ",") <~ "]") | ("[" ~> success(List()) <~ "]")
-
 
   protected lazy val tokenMap: Parser[(Any,Any)] = {
     (token <~ "-" <~ ">") ~ (arrayValues | token) ^^ {
@@ -162,20 +165,22 @@ class XDDdlParser(parseQuery: String => LogicalPlan, xDContext: XDContext) exten
   protected lazy val addJar: Parser[LogicalPlan] =
     ADD ~> JAR ~> restInput ^^ {
       case jarPath =>
-        AddJar(xDContext,jarPath.trim)
+        AddJar(jarPath.trim)
     }
 
-  protected lazy val addApp: Parser[LogicalPlan] =
-    (ADD ~> APP ~> ident) ~ (AS ~> ident).? ~ (WITH ~> className) ^^ {
-      case jarPath ~ alias ~ cname =>
-        AddApp(xDContext, jarPath.toString, cname, alias)
-    }
+
+protected lazy val addApp: Parser[LogicalPlan] =
+  (ADD ~> APP ~> stringLit) ~ (AS ~> ident).? ~ (WITH ~> className) ^^ {
+    case jarPath ~ alias ~ cname =>
+      AddApp(jarPath.toString, cname, alias)
+  }
+
 
   protected lazy val executeApp: Parser[LogicalPlan] =
     (EXECUTE ~> ident) ~ tableValues ~ (OPTIONS ~> options).? ^^ {
       case appName ~ arguments ~ opts =>
         val args=arguments map {arg=> arg.toString}
-        ExecuteApp(xDContext, appName, args, opts)
+        ExecuteApp(appName, args, opts)
     }
   /**
    * Streaming
@@ -305,6 +310,15 @@ class XDDdlParser(parseQuery: String => LogicalPlan, xDContext: XDContext) exten
         }
         Success(streamSql, streamingInfoInput(in))
       }
+    }
+  }
+
+  protected lazy val createGlobalIndex: Parser[LogicalPlan] = {
+
+    CREATE ~ GLOBAL ~ INDEX ~> tableIdentifier ~ (ON ~> tableIdentifier) ~ schemaValues ~ (WITH ~> PK ~> token) ~ (USING ~> className).? ~ (OPTIONS ~> options) ^^ {
+      case index ~ table ~ columns ~ pk ~ provider ~ opts =>
+
+        CreateGlobalIndex(index, table, columns, pk, provider, opts)
     }
   }